replace tracks uikit with swifiui
