+++ MusicAlbums/CollectionViewController.swift
@@ -109,9 +109,8 @@ class CollectionViewController: UICollectionViewController, UICollectionViewDele
     
     override func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
         let album = dataSource.album(at: indexPath)
-//        let viewController = AlbumDetailsViewController(album: album)
         
-        let viewController = UIHostingController(rootView: AlbumDetailsSUIView(album: album))
+        let viewController = UIHostingController(rootView: AlbumDetailsView(album: album))
         viewController.title = album.title
         
         navigationController?.pushViewController(viewController, animated: true)
